bdev/gpt: Remove limitation of 512 multiples when DIF is enabled
If there is metadata and it is interleaved, read data together with
metadata and then strip metadata.

The variable sector_size is used as data block size and ext_sector_size
is used as block size (data + metadata).

Allocate bounce buffer and pass it to spdk_bdev_read if it is allocated,
and in this case use the intermediate callback and in the callback
copy only data from bounce buffer to buffer.

Change-Id: Ia45392955a81328ed2de7d2e28adb41e7a640efe
Signed-off-by: Shuhei Matsumoto <shuhei.matsumoto.xt@hitachi.com>