Update patch set 5
Patch Set 5:

I see the fio and perf tool define these as global options - but for nvme bdev, should they be per-controller options instead?  I think the bdev module should be more configurable.  It is OK for fio and perf to have these as global options - it would be extremely uncommon for someone to want to run perf tests with different options for different controllers.

Patch-set: 5
CC: Gerrit User 1011222 <1011222@d5d70762-12d0-45a1-890d-524b12d3f735>