Update patch set 1
Patch Set 1:

This is an interesting issue. It seems like the port would have to be corrupted in some way for this to happen. It looks like the only times that we free the ports is in spdk_nvmf_rdma_destroy or spdk_nvmf_rdma_stop_listen (which I don't think is actually invoked from anywhere currently.
When do you run into this issue? It seems like it could only occur if we were calling spdk_nvmf_tgt_destroy before the qpairs are all destroyed and disconnected which seems like an error in and of itself.

Patch-set: 1
Reviewer: Gerrit User 1011223 <1011223@d5d70762-12d0-45a1-890d-524b12d3f735>