Update patch set 7
Patch Set 7:

> just not sure about how to get the return value from loading the modules.

As Pawel Kaminski wrote it is good to have *some tests* for this functionality. I don't think we need to invent very sophisticated tests as these macros are mostly for developprs. There are two ways I can think of we can check if these macros are still working putting as little effort as possible:

1. When some SPDK app crashes during tests there can be coredump generated (see test/autotest_common.sh function process_core()) You can put there some commands that will print internal variables. When there be any error in output we can spot it when ispecting failure logs.

2. Some simple test that for excample run SPDK application and do series of gdb commands. If there will be any gbd error in output fail the test.

You can do 1. or 2. or both.

Patch-set: 7