Update patch set 1
Patch Set 1:

This patch is intended to be used when DIF is used locally in a spdk application. If DIF is used in end to end we will not be able to use as per bdev option.  So it may be practical and enough to discard the next three patches and change seed to be a global option

Patch-set: 1
Reviewer: Gerrit User 1012251 <1012251@d5d70762-12d0-45a1-890d-524b12d3f735>