Update patch set 10
Patch Set 10:

This is still just one patch but I'm moved things around to minimize the review effort.  Here's why I did it all in one patch: The prev mem leak was in the init function where we allocated devices in a loop and in the err path only freed the last one allocated - so ASAN passed because right now there's now way for UT to mock an n+1 allocation.  So, in moving all the logic that was in the loop to a function that function then, using the same free(device) call now frees the only device allocated in the function so even on device n+1 the device is freed.  In order to make this 2 patches I'd have to not free the device in the new function which would cause UT to fail w/ASAN because now the first device failure (tested by UT) won't pass.

Make sense?

Patch-set: 10
Reviewer: Gerrit User 1011289 <1011289@d5d70762-12d0-45a1-890d-524b12d3f735>
Work-in-progress: false