Update patch set 1
Patch Set 1:

Hi Shuhei,

I see the issue in spdk_nvmf_tcp_qpair_destroy when the transport recycling the resource. Previously, we did not fix the issue in perf, so the perf continues polling. And our target did not handle this well. So it is the issue on the initiator, but I think that we still need to add this check to make our NVMe-oF target robust.

Patch-set: 1
Reviewer: Gerrit User 1011275 <1011275@d5d70762-12d0-45a1-890d-524b12d3f735>