Update patch set 3
Patch Set 3:

I added UT code to verify guard is zero if block is all zero and seed is zero, and guard is not zero if block is all zero but seed is not zero. I have added seed to most test cases and this addition will be enough.

Patch-set: 3