Update patch set 2
Patch Set 2:

Hi Shuhei,

I revised code. Besides this one, I think that there are other spaces which only limit the format to Decimal. It seems that it is not quite friendly. 

But for here, I think that we should support to pase 0, to enable hex format especially for staring app with dpdk

Patch-set: 2