bdev: Add spdk_bdev_get_data_block_size API.
When iSCSI target supports DIF transparently, iSCSI initiator is not
aware of DIF and it assumes the block device don't have any metadata.

iSCSI target is already matured and the major use case of DIF will
be that iSCSI target inserts and strips DIF on behalf of iSCSI initiator.

We focus on such a use case in SPDK iSCSI target and SCSI, and need
that SCSI layer processes SCSI command without metadata.

This patch adds an new API spdk_bdev_get_data_block_size to bdev
for that case.

Change-Id: I0fd2a6d0bcf6a4c18c583f70d96cc5035fc57fe9
Signed-off-by: Shuhei Matsumoto <shuhei.matsumoto.xt@hitachi.com>