iscsi: Make the data out buffer size configurable by JSON RPC.
To insert DIF into read data in data out buffer, we have to
make the size of data out buffer larger than 64KB.

Hence this patch makes the data out buffer size variable and
configurable by JSON RPC.

Set the minimum and maximum size of the data out buffer size
to 65536 and (65536 * 2), respectively because it needs
to store DIF when the reported maximum size is 65536.

The existing format which has the largest portion of metadata,
is 16 byte metadata per 512 byte block. If we use the format,
the necessary size is the following.

  (65536 / 512) * (512 + 16) = 67584.

Hence 67584 may be better than 65536 * 2 as the maximum size.

Change-Id: I215e95ccc5058d1380b99a8d5a2439cac32ac361
Signed-off-by: Shuhei Matsumoto <shuhei.matsumoto.xt@hitachi.com>