Update patch set 10
Patch Set 10:

Thanks Shuhei but I've spent about as much time trying to refactor this for review-ability as I did fixing it to being with :) I revived with Jim and showed him the subtle issues with trying to make one patch that just moves the code and another that fixes it and he's OK with it "as is" right now.

If you like though, I'll take a look at your suggestions and give it another shot though. Let me know...

Patch-set: 10
Reviewer: Gerrit User 1011289 <1011289@d5d70762-12d0-45a1-890d-524b12d3f735>