Update patch set 10
Patch Set 10:

> Patch Set 10:
> 
> Hi Paul, Jim's suggestion makes sense to me. I just thought that fixing memory leak may have to be done first. Please go forward with Jim's suggestion. I of course agree with that.

Haha, so Jim's suggestion is exactly what I was trying to do.  Let me know if you want me to explain exactly why it was more challenging than it was worth.  Jim is OK with it as is, I did clean it up a bit so there are minimal changes in the final review

Patch-set: 10
Reviewer: Gerrit User 1011289 <1011289@d5d70762-12d0-45a1-890d-524b12d3f735>