Update patch set 2
Patch Set 2: Code-Review-1

Thanks for the explanation Ziye. The traceback is very useful.
I think that there may be a more general issue with the qpair disconnect path in TCP. Specifically, we are risking not freeing buffers back to the poll group buffer pool.
If you look at the RDMA code, there is no poll_group_remove function in the function table. But there is one in the TCP code.
In spdk_nvmf_tcp_poll_group_remove, you set tqpair->group = NULL. This results in the following chain of events during qpair disconnect (starting from step 8 in your traceback below):
1. _spdk_nvmf_qpair_destroy calls
2. spdk_nvmf_tcp_poll_group_remove (sets qpair->group = NULL)
3. spdk_nvmf_transport_qpair_fini
4. spdk_nvmf_tcp_close_qpair
5. spdk_nvmf_tcp_cleanup_all_states (causes all requests to be freed)
6. nvmf_tcp_request_free
7. spdk_nvmf_tcp_request_free_buffers (for any request that reports req->data_from_pool = true).
8. The qpair->group has already been set to NULL which results in a segfault in the current implementation. In your change, it causes the function to bail without freeing any buffers associated with that I/O.

Not freeing the buffers is O.K. in the path where we are shutting down, but in the scenario where we are only closing one qpair, we could accidentally permanently steal buffers from the data_buf_pool.

I think that the better solution here is to not set rqpair->group = NULL in spdk_nvmf_tcp_poll_group_remove. That allows us to still free the request resources back to the group on shutdown. Plus, since the lifecycle of the group is longer than that of any qpairs, we know that if the qpair still exists the group will not have been freed, making it safe to keep the group pointer active. We do similar things in the RDMA transport.

Patch-set: 2
Reviewer: Gerrit User 1011223 <1011223@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1