rdma: optimize the request state machine.
The state machine sometimes jumps back into the switch statement when it
doesn't have to, and it would break after every state even when
fallthrough is possible and more efficient. A recent performance
benchmark showed us spending up to 10% of CPU time in
spdk_nvmf_rdma_request_process in NVMe-oF workloads. This patch is aimed
at addressing some of that overhead.
see https://github.com/spdk/spdk/issues/669 for the perf report.

Change-Id: I230358f9c8254595568059a413294c0049d3d48c
Signed-off-by: Seth Howell <seth.howell@intel.com>