nvme/qpair: remove the if judgement.
It seems that this check is too strong.

Reason: If the users submit large I/O size which exceeds the
provided max transfer size, the I/O will be split. For
example, if we create I/O depth = 128, and the max transfer size
is 128KB. Then if the users do I/O size with 512 KB, each
NVMe req will be splitted into 5 (1 original + 4 req), thus
there will be 512 read qd, which will trigger the waiting policy
and put it into the queued_req list. And it will be selected to
submit via call nvme_qpair_submit_request again.

So waiting for the submission is the policy defined to queued the
I/O for handling large I/O size. I have concern that if
we add the check here, we will fail many I/Os in such case.

Change-Id: I0c682161d298f97b327c9dfbf020f85adc143e18
Signed-off-by: Ziye Yang <ziye.yang@intel.com>