nvmf: use master thread when destroying target.
I discovered an issue when investigating issue 682. We were hitting an
assert in spdk_bdev_close when trying to close a bdev that had been
created via RPC. This occurred because the bdevs are created from the
master thread (where the RPC server is running) and we were forcing the
target to destroy from the first core in the core mask, not necessarily
where the master thread is.

This change ensures that the master thread lines up with the thread on
which we destroy the target.

Change-Id: I00323c94cd7f6fb4665462b36c33b53222e272ea
Signed-off-by: Seth Howell <seth.howell@intel.com>