scripts/rpc: Change the separator of portal list parameter of add_portal_group
The python script for add_portal_group had required each portal is
an argument and portal list are made of remaining all arguments.

However this prevents adding new parameters to add_portal_group.

This patch changes portal list to be a single argument for python,
and each portal is separated by a whitespace in the argument.

Currently there will be very rare that a portal group has multiple
portals. Hence this change will not break any compatibility.

Change-Id: I8c597a1194fff338afb371e6d24f9273122b753c
Signed-off-by: Shuhei Matsumoto <shuhei.matsumoto.xt@hitachi.com>