iscsi: Add an option to enable DIF insert and strip as of portal group
Subsequent patches will support DIF insert for write I/O and DIF strip
for read I/O in iSCSI target. This feature is intended to be end-to-end,
and DIF processing will be done in iSCSI target next to the network
socket. Hence adding an option to enable the feature as per portal group
option and copying it to the corresponding connections at creation
will be reasonable. The next patch will set the option of the connection
at its creation.

Change-Id: I8f39b9e1045f5f8e05b288875ec9f8fafdf121fd
Signed-off-by: Shuhei Matsumoto <shuhei.matsumoto.xt@hitachi.com>