Update patch set 9
Patch Set 9: Code-Review+1

This is really cool! It hadn't even occurred to me that the signal handler would be signaled on an arbitrary thread as opposed to the thread that it was registered on, but sure enough, it looks like that is exactly what can happen.

With this in mind, I think we have a race condition between subsystem_fini_next and subsystem_init_next, as follows:
initial state: g_subsystems_initialized == false
1. subsystem_init calls g_next_subsystem = TAILQ_NEXT resulting in g_next_subsystem = NULL
2. subsystem_fini executes if (!g_next_subsystem) takes true branch
3. subsystem_fini checks g_subsystems_initialized and finds it is false.
4. subsystem_init sets g_subsystems_initialized = true, but it is too late.
5. subsystem_fini exits without actually iterating through the subsystems and we end up with memory leaks.

You didn't introduce this race in your patch, but I think that it's a real one, and we should introduce a patch on top of yours with a lock for init/fini of subsystems to completely fix the shutdown path.

Patch-set: 9
Reviewer: Gerrit User 1011223 <1011223@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=+1