Update patch set 36
Patch Set 36:

We've talked quite a bit on this now and I'm still struggling to understand a number of things. To try and make some progress, I'll outline what I think needs to be happening here, and then maybe you can fill me in on which part I'm missing. I suspect my confusion on this patch is that I'm expecting it to be one thing and it's actually for an entirely different purpose.

At a very high level, we need some mechanism whereby our libraries can register event types with some global registry. There should then be an RPC supported that lists which event types exist. Clients (i.e. Python processes) should be able to query the list of available event types and then send an RPC to subscribe to the event. Whenever the event occurs, the client then gets a message with the relevant details. Is this close to what you guys are building, or am I way off?

At an implementation level there are a couple of challenges to make this scale up properly. The first challenge is that the RPC server can't just send unsolicited messages to the clients - it has to have an open socket and a request id in order to generate a response. I think we deal with this by making the clients send a JSON-RPC subscribe message which does not get a response until the event occurs. The event is essentially the response to the subscribe message, which also holds the socket open. These subscriptions would be one time use - to get new events after one fires the client just sends the subscribe message again.

The second challenge is scaling the broadcast of an event in the case where there are lots of clients subscribed. This needs to occur asynchronously. In the most efficient scenario, we'd generate the relevant JSON for the event one time and then asynchronously try to send it to all of the subscribed clients until they all have received it. This can be done entirely internal to the notification library via a poller - the caller of spdk_notify_send or whatever we call it does not need to be aware of these details.

Patch-set: 36
Reviewer: Gerrit User 1010531 <1010531@d5d70762-12d0-45a1-890d-524b12d3f735>