Update patch set 37
Patch Set 37:

> Patch Set 37:
> 
> Hi Piotr and Pawel,
> 
> Did you consider taking an approach similar to NVMe AER?  Where the notification is very minimal - just enough information that the client knows how to figure out what changed?  It seems like that would significantly reduce both the complexity and the size of this feature.
> 
> -Jim

I think that this is minimal solution for notifications in spdk.
The complications, that Ben mentioned in his review solve technical issues and are driven by requirements that we discussed during the architecture of the feature.
- List of available types of notifications. This introduces two structures instead of one. 
- The way of subscribing to notifications is different at library level and RPC level to reconcile minimal approach and json-rpc requirements

I think that Ben perfectly summarized problems we face in this feature in his comment.

Patch-set: 37
Reviewer: Gerrit User 1011211 <1011211@d5d70762-12d0-45a1-890d-524b12d3f735>