Update patch set 11
Patch Set 11:

> Patch Set 11:
> 
> (1 comment)

The intent of this change is to optimize for performance over storage utilization.  By repeatedly generating the tables (especially for test suites that run a large number of small, different iterations of individual tests) this would be excessive and the original implementation (which effectively did all of the calculations “real time”) would be just as good.

Patch-set: 11
Reviewer: Gerrit User 1015860 <1015860@d5d70762-12d0-45a1-890d-524b12d3f735>