iscsi: Use buffer length as length of an iovec for write I/O
Previously data length was used as length of an iovec for write I/O.
However, if we support DIF insert/strip feature, the actual data
length sent to SCSI layer will be larger than the data length set
in the iSCSI PDU. Hence this patch changes to use buffer length
as length of an iovec for write I/O. SCSI task holds data length
as an independent member and use it. Hence this change doesn't lose
data length and causes no harm.

Change-Id: Iacf51b6e932dd06dbc26824ef27cb87b4a9da760
Signed-off-by: Shuhei Matsumoto <shuhei.matsumoto.xt@hitachi.com>